Added a formattable toString function.
This feature adds a version of the toString function that allows to use
a processor class to generate a formatted string output from either the
simple class name or a custom name and a mapping of members and their
values. This allows for example to append the type literals like d, f or
l to the values. It also allows for example to change the representation
of floating point numbers. For providing the new toString function a new
interface FormattableToString was created, which requires at least to
return a mapping of the members and values or null. As for the processor
the interface ToStringFormatter was created. The default implementation
of it appends the type literals like d, f or l to the values. Also it
formats any value which type implements the FormattableToString
interface with the default ToStringFormatter defined for that type.

- Added the interface FormattableToString to the API as the target for
formatting.
- Added the interface ToStringFormatter to the API as the processor of
the formatting.
- Added a default implementation for ToStringFormatter.
- Implemented the FormattableToString interface in all tuple classes.
- Added a new general property for the default ToStringFormatter in the
Barghos class.