Added an event bus system additionally to the existing event system.
An event bus system was added, that differs from the existing .NET style
event system in being a broadcast system for events through an entire
program or parts of it. It works by creating an event bus and
subscribing an event bus subscriber to it. The event bus subscriber is
either a class that implements the EventBusSubscriber interface and
therefore registers it's event handler methods manually or a class with
the EventBus.Subscriber annotation. In the latter case all the event
handler methods have to have the EventBus.EventHandler annotation with
the event type specified. Any event handler can be assigned to multiple
events as long as the parameter of the event handler can accept all
event types in question.

The package structure has been changed as some classes are actually
required by the API and are therefore part of it.

A reflection helper for classes has been added as additional
functionality for retrieving annotated methods and functions with a
specific annotation from a class was required.

A collection utillity class was added to provide more functionality for
operations with all kind of collections. It currently adds a contains
function that accepts the functional interface "Function" and therefore
allows to scan a collection by calling the given Function implementation
for each entry. Therefore it is now possible to scan a list of instances
for a specific value in one of their values.

A value relay has been added for easier testing of calls to anonymous
implementations.