Added validation functions to tuple readonly interfaces.
- Added a function isValid to all tuple readonly interfaces to provide
an easy way to validate the data in a tuple. By default the BigDecimal,
the BigInteger, the Object, the generic and the String-tuples are only
valid if none of the components is null.

- Added the function isFinite to all numeric tuple readonly interfaces.
By default the isFinite function returns true for all numeric tuples
except the float and double tuples as these types deal with infinity and
NaN. For these tuples the corresponding function isFinite like
Double.isFinite is called for every component. A tuple is only
considered to be finite if all components are finite.

- Added the function isZero to all numeric tuples readonly interfaces.
It returns true if all components are exactly zero.

- Added the alternative function isZero(with tolerance) to all numeric
readonly interfaces. It returns true if all components are within the
tolerance around zero, positive and negative. For example, for a
tolerance of 0.5 a value of 0.4 is considered to be zero, a value of
-0.4 aswell but a value of 0.6 or -0.6 is not considered to be zero.