Added getNewInstance functions to tuples for type continuity.
Added three new functions getNewInstance to the readonly interfaces of
all tuple classes. The three functions handle the default inputs of
either a single tuple of the same data type, a single value which is
used for all components or dedicated parameters for each component. The
functions with the tuple and value parameter have a default
implementation direct in the readonly interfaces. They call the function
with the dedicated Parameters. These functions allow to create a new
instance of the original type. With this there is no need to know the
exact type for for example a clone operation. The desired behaviour of
these functions is always to return a new instance of the original type.
That means that if a function gets a parameter of the type Tup3fR passed
and the original type was a PTup3f, then if this function calls the
getNewInstance function the returned value will still be reduced to
Tup3fR but the created type will again be a PTup3f. This behaviour also
allows cloning actions in a default implementation of functions, where
before was always the question what type to return. From a logical point
this is also very important, as the behavior of any implemented function
can vary from type to type.