Extended the tuple-interfaces by N-Tuple layer.
- There is now a new base layer that is shared by all tuples
representing an N-Tuple. Therefore it provides functionality for
array-like access to components.
- The tuple-interfaces where extended by validity functions for control
over the state of the object by the implementation.
- The tuple-interfaces where extended by functions for abstract type
instantiation. With this it is now possible from any abstraction layer
of the tuple to generate a new instance of the same original type. This
is useful for example for the default functions in the readonly
interfaces, that need to return a modified copy of the original.