Optimized tuple indices.
- Removed the tuple component index constants from the readonly
interfaces and added them in a single new class TupleConstants. With
this the constants are created once for all tuples.
- Replaced direct index references (i.e. v[1]) in the tuple interfaces
with the new component index constants. It should be less likely now to
mess up the indices and also it is better understandable.
- Removed some unnecessary variables from the setArray and set(TupR)
functions in the tuple interfaces.
- Removed automatic build scans as they need manual permission for
publishing and therefore keep build processes running.